requirements.txt
pyserial~=3.5
mido~=1.3.3
smbus2~=0.5.0
requests~=2.32.3
python-rtmidi==1.5.8
